% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addAnalysisVariables}
\alias{addAnalysisVariables}
\title{Add demographic variables needed for analysis}
\usage{
addAnalysisVariables(geMetaData)
}
\arguments{
\item{geMetaData}{single data.table containing all gene expression meta-data}
}
\description{
Add demographic variables needed for analysis
}
