% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{addArmAccession}
\alias{addArmAccession}
\title{Add Arm Accession to meta-data}
\usage{
addArmAccession(dt, gef)
}
\arguments{
\item{dt}{meta-data data.table}

\item{gef}{gene expression files table from ImmuneSpace connection}
}
\description{
Add Arm Accession to meta-data
}
