% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addCoalescedFeatureSetName}
\alias{addCoalescedFeatureSetName}
\title{Create new featureSetName field that coalesces similar platforms for analysis}
\usage{
addCoalescedFeatureSetName(dt)
}
\arguments{
\item{dt}{data table}
}
\description{
Create new featureSetName field that coalesces similar platforms for analysis
}
