% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponsePreProcessing.R
\name{addDemographicFields}
\alias{addDemographicFields}
\title{Add fields from demographic data to meta-data}
\usage{
addDemographicFields(dt, demographicData)
}
\arguments{
\item{dt}{meta-data data.table}

\item{demographicData}{demographic data data.table}
}
\description{
Add fields from demographic data to meta-data
}
