% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addFeatureAnnotationSetName}
\alias{addFeatureAnnotationSetName}
\title{Add name of feature annotation set name used to generate gene expression matrix}
\usage{
addFeatureAnnotationSetName(geMetaData, featureAnnotationMap)
}
\arguments{
\item{geMetaData}{single data.table containing all gene expression meta-data}

\item{featureAnnotationMap}{feature annotation set map table from ImmuneSpace DB}
}
\description{
Add name of feature annotation set name used to generate gene expression matrix
}
