% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addFeatureAnnotationSetVendor}
\alias{addFeatureAnnotationSetVendor}
\title{Add name of feature annotation set vendor used to generate gene expression matrix}
\usage{
addFeatureAnnotationSetVendor(geMetaData, featureAnnotation)
}
\arguments{
\item{geMetaData}{single data.table containing all gene expression meta-data}

\item{fas}{feature annotation set table from ImmuneSpace DB}
}
\description{
Add name of feature annotation set vendor used to generate gene expression matrix
}
