% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{addGeBatchName}
\alias{addGeBatchName}
\title{Add geBatchName for use in cross-study normalization}
\usage{
addGeBatchName(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Add geBatchName for use in cross-study normalization
}
