% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addGSMAccessions}
\alias{addGSMAccessions}
\title{Add name of feature annotation set vendor used to generate gene expression matrix}
\usage{
addGSMAccessions(geMetaData, gef)
}
\arguments{
\item{geMetaData}{single data.table containing all gene expression meta-data}

\item{gef}{gene expression files dataset from ImmuneSpace connection object}
}
\description{
Add name of feature annotation set vendor used to generate gene expression matrix
}
