% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addImmuneExposureFields}
\alias{addImmuneExposureFields}
\title{Add fields from immune exposure data to meta-data}
\usage{
addImmuneExposureFields(dt, immuneExposureData)
}
\arguments{
\item{dt}{meta-data data.table}

\item{immuneExposureData}{immune exposure data.table}
}
\description{
Add fields from immune exposure data to meta-data
}
