% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{addIrpBatchName}
\alias{addIrpBatchName}
\title{Add irpBatchName for use in responseCall generation}
\usage{
addIrpBatchName(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Add irpBatchName for use in responseCall generation
}
