% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addMatrixRelatedFields}
\alias{addMatrixRelatedFields}
\title{Add matrix, featureset, and cell_type fields to pData objects}
\usage{
addMatrixRelatedFields(phenoDataSets, geMatrices)
}
\arguments{
\item{phenoDataSets}{list of pData objects}

\item{geMatrices}{ImmuneSpace connection con$cache$GE_matrices object}
}
\description{
Add matrix, featureset, and cell_type fields to pData objects
}
