% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructWithResponseEsets.R
\name{addResponseData}
\alias{addResponseData}
\title{Add selected immune response data to final expressionSets}
\usage{
addResponseData(ageCohort, noResponseEsets, selectedImmdata)
}
\arguments{
\item{ageCohort}{name of age cohort}

\item{noResponseEsets}{expressionSets without response data}

\item{selectedImmdata}{selected immune response data}
}
\description{
Add selected immune response data to final expressionSets
}
