% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{addSDY1325metadata}
\alias{addSDY1325metadata}
\title{Add fields from SDY1325 custom meta-data to shared meta-data}
\usage{
addSDY1325metadata(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Add fields from SDY1325 custom meta-data to shared meta-data
}
