% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{addTimePostLastVax}
\alias{addTimePostLastVax}
\title{To accomodate studies with booster shots, add a timepoint post final vaccination}
\usage{
addTimePostLastVax(dt)
}
\arguments{
\item{dt}{pData from expressionSet object}
}
\description{
To accomodate studies with booster shots, add a timepoint post final vaccination
}
