% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponsePreProcessing.R
\name{addVaccineFields}
\alias{addVaccineFields}
\title{Add fields from vaccine data to meta-data}
\usage{
addVaccineFields(dt, vaccines)
}
\arguments{
\item{dt}{meta-data data.table}

\item{vaccineData}{vaccine data data.table}
}
\description{
Add fields from vaccine data to meta-data
}
