% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{adjustProblemSamples}
\alias{adjustProblemSamples}
\title{Remove selected participants from expression set}
\usage{
adjustProblemSamples(eset, problemSamples)
}
\arguments{
\item{eset}{expressionSet}

\item{problemSamples}{vector of participant IDs to flag as failedGenderQC}
}
\description{
Remove selected participants from expression set
}
