% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{adjustProblemStudies}
\alias{adjustProblemStudies}
\title{Remove selected participants from expression set}
\usage{
adjustProblemStudies(eset, problemStudies)
}
\arguments{
\item{eset}{expressionSet}

\item{participantIdsToRm}{vector of participant IDs to remove}
}
\description{
Remove selected participants from expression set
}
