% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageGroups}
\alias{ageGroups}
\title{Age groups for determining age cohorts}
\format{A data frame with three variables: \code{count}, \code{study},
  and \code{age_bracket}}
\usage{
ageGroups
}
\description{
Data created from inputSamples query in ImmuneSpace. Age_bracket variable
is the ceiling of that decile.
}
\keyword{datasets}
