% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_genesets}
\alias{all_genesets}
\title{Gene sets used in the pre-vaccination manuscript}
\format{
A list of lists
}
\usage{
all_genesets
}
\description{
Genesets combining the hallmark gene set and blood transcription modules
}
\keyword{datasets}
