% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{batchCorrect.importedModel}
\alias{batchCorrect.importedModel}
\title{Adjust baseline expression data to account for study_accession, featureSetName, and cell_type}
\usage{
batchCorrect.importedModel(modelEset, targetEset, batch.vars, covariates)
}
\arguments{
\item{modelEset}{expressionSet used to create linear model}

\item{targetEset}{expressionSet that is corrected given model fit values}
}
\description{
Adjust baseline expression data to account for study_accession, featureSetName, and cell_type
}
