% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{batchCorrect}
\alias{batchCorrect}
\title{Adjust expression data using correction coefficients based on baseline data
to account for study_accession, featureSetName, and cell_type}
\usage{
batchCorrect(eset, batch.vars, covariates)
}
\arguments{
\item{eset}{expressionSet}

\item{batch.vars}{variables to correct for}

\item{covariates}{additional variables to include in the model}
}
\description{
Adjust expression data using correction coefficients based on baseline data
to account for study_accession, featureSetName, and cell_type
}
