% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{batchCorrectBaselineData}
\alias{batchCorrectBaselineData}
\title{Adjust baseline expression data to account for study_accession, featureSetName, and cell_type}
\usage{
batchCorrectBaselineData(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Adjust baseline expression data to account for study_accession, featureSetName, and cell_type
}
