% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFinalEsets.R
\name{checkFinalTestResults}
\alias{checkFinalTestResults}
\title{Ensure eset is as expected}
\usage{
checkFinalTestResults(results)
}
\arguments{
\item{results}{test results list}
}
\description{
Ensure eset is as expected
}
