% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{correctHrs}
\alias{correctHrs}
\title{Convert 'Hours'-based study times to 'Days'-based}
\usage{
correctHrs(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Convert 'Hours'-based study times to 'Days'-based
}
