% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createUniqueIdColumn}
\alias{createUniqueIdColumn}
\title{Generate a unique id column}
\usage{
createUniqueIdColumn(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Generate a unique id column
}
