% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{crossStudyNormalize}
\alias{crossStudyNormalize}
\title{Cross-Study normalize an expressionSet object based on a single
vendor's subset of expression data to define the target distribution}
\usage{
crossStudyNormalize(eset, vendorToUse, studiesToExclude)
}
\arguments{
\item{eset}{expressionSet}

\item{vendorToUse}{Microarray or RNAseq vendor to use for generating target distribution}

\item{studiesToExclude}{studies to exclude when generating the target distribution}
}
\description{
Cross-Study normalize an expressionSet object based on a single
vendor's subset of expression data to define the target distribution
}
