% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{customProcessing}
\alias{customProcessing}
\title{Custom pre-processing for particularly ornery studies}
\usage{
customProcessing(assay, df)
}
\arguments{
\item{assay}{assay name}

\item{df}{assay data-frame}
}
\description{
Custom pre-processing for particularly ornery studies
}
