% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{elisaResponseCall}
\alias{elisaResponseCall}
\title{Generate immune response call for age specific ELISA cohort}
\usage{
elisaResponseCall(
  dt,
  discretizationValues,
  ageCutoff,
  postVaxDayRange,
  isYoung
)
}
\arguments{
\item{dt}{immdata df}

\item{discretizationValues}{points to use for cutting of low, mid, high responders}

\item{ageCutoff}{age at which young and older are defined}

\item{isYoung}{boolean}
}
\description{
Generate immune response call for age specific ELISA cohort
}
