% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expectedGeMetaDataColumns}
\alias{expectedGeMetaDataColumns}
\title{Expected Gene Expression Meta-Data columns}
\format{
string vector containing all column names
}
\usage{
expectedGeMetaDataColumns
}
\description{
After pre-processing, a vector of column names expected to be found in geMetaData
}
\keyword{datasets}
