% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFinalEsets.R
\name{filterEsetByAge}
\alias{filterEsetByAge}
\title{Filter expressionSet by age cutoffs}
\usage{
filterEsetByAge(noNormEset, ages)
}
\arguments{
\item{noNormEset}{expressionSets without response data}

\item{ages}{allowed ages, either one or two sets of low and high points}
}
\description{
Filter expressionSet by age cutoffs
}
