% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponsePreProcessing.R
\name{filterImmdataByAge}
\alias{filterImmdataByAge}
\title{Filter immdata list elements by age filter}
\usage{
filterImmdataByAge(immdata, ageCutoff, isYoung)
}
\arguments{
\item{immdata}{list of assay data.table(s)}

\item{ageCutoff}{integer value for cutting age_imputed}

\item{isYoung}{boolean selecting young or older cohort}
}
\description{
Filter immdata list elements by age filter
}
