% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{filterOutNoVaccineSamples}
\alias{filterOutNoVaccineSamples}
\title{Filter out cohorts with no vaccine data from custom mapping}
\usage{
filterOutNoVaccineSamples(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Filter out cohorts with no vaccine data from custom mapping
}
