% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{generateELISAResponse}
\alias{generateELISAResponse}
\title{Generate immune response callELISA}
\usage{
generateELISAResponse(dt)
}
\arguments{
\item{dt}{immdata df}
}
\description{
Generate immune response callELISA
}
