% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{generateResponseCall}
\alias{generateResponseCall}
\title{Generate immune response calls for HAI, NAb or ELISA}
\usage{
generateResponseCall(assay, data, postVaxDayRange, discretizationValues)
}
\arguments{
\item{assay}{assay name}

\item{data}{immdata df}

\item{postVaxDayRange}{Allowable timepoints for post-vaccine values}

\item{discretizationValues}{cut points to use for discretizing response call groups}
}
\description{
Generate immune response calls for HAI, NAb or ELISA
}
