% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{generateSampleMDSPlot}
\alias{generateSampleMDSPlot}
\title{Generate sample MDS plots for QC}
\usage{
generateSampleMDSPlot(eset, numberOfSamples)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate sample MDS plots for QC
}
