% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponsePreProcessing.R
\name{getImmuneResponseData}
\alias{getImmuneResponseData}
\title{Function generates a base dataset from ImmuneSpace}
\usage{
getImmuneResponseData(assay, con, studies)
}
\arguments{
\item{assay}{assay name in ImmuneSpace}

\item{con}{ImmuneSpaceR connection object}

\item{studies}{list of ImmuneSpace studies to use in filtering}
}
\description{
Function generates a base dataset from ImmuneSpace
}
