% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{getLmFitValues}
\alias{getLmFitValues}
\title{Get linear model fit values for each gene}
\usage{
getLmFitValues(model.formula, eset.baseline, coefs2adjust)
}
\arguments{
\item{model.formula}{model formula}

\item{eset.baseline}{baseline expressionSet}
}
\description{
Get linear model fit values for each gene
}
