% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{getModelFormula}
\alias{getModelFormula}
\title{Get linear model formula}
\usage{
getModelFormula(model.vars)
}
\arguments{
\item{model.vars}{vector of variables}
}
\description{
Get linear model formula
}
