% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getTable}
\alias{getTable}
\title{Get a backend table not easily accessible through ImmuneSpaceR}
\usage{
getTable(con, schemaName, queryName, ...)
}
\arguments{
\item{con}{ImmuneSpaceR connection object}

\item{schemaName}{schema name in ImmuneSpace DB}

\item{queryName}{query name in ImmuneSpace DB}

\item{...}{additional arguments passed to labkey.selectRows.}
}
\description{
Get a backend table not easily accessible through ImmuneSpaceR
}
