% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedMetaData.R
\name{imputeAge}
\alias{imputeAge}
\title{Impute age for those with missing age values or ranges given by ImmPort}
\usage{
imputeAge(dt)
}
\arguments{
\item{dt}{meta-data data.table}
}
\description{
Impute age for those with missing age values or ranges given by ImmPort
}
