% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{imputeGender}
\alias{imputeGender}
\title{Impute gender based on expression of Y chromosome-related genes}
\usage{
imputeGender(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Impute gender based on expression of Y chromosome-related genes
}
