% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{imputeGender.useAllTimepoints}
\alias{imputeGender.useAllTimepoints}
\title{Impute gender based on expression of Y chromosome-related genes}
\usage{
imputeGender.useAllTimepoints(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Impute gender based on expression of Y chromosome-related genes
}
