% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{imputeYchrom.useAllTimepoints}
\alias{imputeYchrom.useAllTimepoints}
\title{Impute presence of y chromosome based on expression of Y chromosome-related genes
and flag for QC if values do not match}
\usage{
imputeYchrom.useAllTimepoints(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Impute presence of y chromosome based on expression of Y chromosome-related genes
and flag for QC if values do not match
}
