% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{performTiterAnalysis}
\alias{performTiterAnalysis}
\title{Flexible function that wraps around `titer::Calculate` methods while
also handling study * age_cohorts that cannot be modelled with an exponential model}
\usage{
performTiterAnalysis(titer_list, analysis, discretize)
}
\arguments{
\item{titer_list}{pre-processed list of dataframes with hai or nab data}

\item{analysis}{RBA or MFC}

\item{discretize}{cut points to use for discretizing response call groups}
}
\description{
Flexible function that wraps around `titer::Calculate` methods while
also handling study * age_cohorts that cannot be modelled with an exponential model
}
