% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{preProcessImmData}
\alias{preProcessImmData}
\title{Pre-processing for 'titer::Calculate' functions}
\usage{
preProcessImmData(dt, postVaxDayRange)
}
\arguments{
\item{dt}{data.table with HAI or NAb data}

\item{postVaxDayRange}{range of integer values for study_time_collected cutpoints}
}
\description{
Pre-processing for 'titer::Calculate' functions
}
