% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.createSubjectsByStudyTable}
\alias{qualityControl.createSubjectsByStudyTable}
\title{Generate table of studies with count of participants from vector of pids}
\usage{
qualityControl.createSubjectsByStudyTable(participantIds)
}
\arguments{
\item{participantIds}{participant IDs}
}
\description{
Generate table of studies with count of participants from vector of pids
}
