% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.failedGenderImputation}
\alias{qualityControl.failedGenderImputation}
\title{Generate gender box plots after QC}
\usage{
qualityControl.failedGenderImputation(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate gender box plots after QC
}
