% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.failedYchromQC}
\alias{qualityControl.failedYchromQC}
\title{Generate ychrom box plots after QC}
\usage{
qualityControl.failedYchromQC(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate ychrom box plots after QC
}
