% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.genderByMatrix}
\alias{qualityControl.genderByMatrix}
\title{Generate sample MDS plots for QC}
\usage{
qualityControl.genderByMatrix(eset, returnObject = "allMatricesPlot")
}
\arguments{
\item{eset}{expressionSet}

\item{returnObject}{options are allMatricesPlot (default), probSamplesPlot, probSamplesDT}
}
\description{
Generate sample MDS plots for QC
}
