% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.sampleMDSPlot}
\alias{qualityControl.sampleMDSPlot}
\title{Generate sample MDS plots for QC}
\usage{
qualityControl.sampleMDSPlot(eset, numberOfSamples = 10)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate sample MDS plots for QC
}
