% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.samplePlot}
\alias{qualityControl.samplePlot}
\title{Generate sample MDS plots for QC}
\usage{
qualityControl.samplePlot(
  eset,
  method = "MDS",
  numberOfSamples = 10,
  colorCol = "study_accession"
)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate sample MDS plots for QC
}
