% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.yChromPresentByMatrix}
\alias{qualityControl.yChromPresentByMatrix}
\title{Generate box plots showing ychrom imputed after QC}
\usage{
qualityControl.yChromPresentByMatrix(eset, returnObject = "allMatricesPlot")
}
\arguments{
\item{eset}{expressionSet}

\item{returnObject}{options are allMatricesPlot (default), probSamplesPlot, probSamplesDT}
}
\description{
Generate box plots showing ychrom imputed after QC
}
