% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{relevelEsets}
\alias{relevelEsets}
\title{Get linear model fit values for each gene}
\usage{
relevelEsets(esets, factorsToRelevel)
}
\arguments{
\item{esets}{expressionSets}

\item{factorsToRelevel}{factors to relevel}
}
\description{
Get linear model fit values for each gene
}
