% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{removeSDY212MissingSample}
\alias{removeSDY212MissingSample}
\title{SDY212 has a biosample that does not have matching meta-data and cannot
be explained by the study authors. This function removes that sample.}
\usage{
removeSDY212MissingSample(esets)
}
\arguments{
\item{esets}{list of expressionSet objects}
}
\description{
SDY212 has a biosample that does not have matching meta-data and cannot
be explained by the study authors. This function removes that sample.
}
