% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{removeSubjectsWithoutBaseline}
\alias{removeSubjectsWithoutBaseline}
\title{Ensure every subject has baseline gene expression data}
\usage{
removeSubjectsWithoutBaseline(eset)
}
\arguments{
\item{eset}{expressionSet object}
}
\description{
Ensure every subject has baseline gene expression data
}
