% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{removeSubjectsWithoutGenderConsensus}
\alias{removeSubjectsWithoutGenderConsensus}
\title{Remove selected participants from expression set}
\usage{
removeSubjectsWithoutGenderConsensus(eset, participantIdsToRm)
}
\arguments{
\item{eset}{expressionSet}

\item{participantIdsToRm}{vector of participant IDs to remove}
}
\description{
Remove selected participants from expression set
}
