% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{removeTimepointFromEset}
\alias{removeTimepointFromEset}
\title{Remove samples from certain timepoints from desired studies}
\usage{
removeTimepointFromEset(esets, study, timepoint)
}
\arguments{
\item{esets}{list of expressionSetObjects}

\item{study}{study id}

\item{timepoint}{study_time_collected value}
}
\description{
Remove samples from certain timepoints from desired studies
}
