% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{runAllAnalyses}
\alias{runAllAnalyses}
\title{run both maxRBA and MFC calculations using the 'titer' package by
Stefan Avey of Yale University}
\usage{
runAllAnalyses(titer_list, df, discretizationValues)
}
\arguments{
\item{titer_list}{pre-processed list of dataframes with hai or nab data}

\item{df}{immdata df}

\item{discretizationValues}{cut points to use for discretizing response call groups}
}
\description{
run both maxRBA and MFC calculations using the 'titer' package by
Stefan Avey of Yale University
}
