% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdy1325_metadata}
\alias{sdy1325_metadata}
\title{SDY1325 meta-data
Derived from study-associated paper's figure 1.}
\format{
data.table
}
\usage{
sdy1325_metadata
}
\description{
SDY1325 meta-data
Derived from study-associated paper's figure 1.
}
\keyword{datasets}
