% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdy1370_elisa}
\alias{sdy1370_elisa}
\title{SDY1370 ELISA data}
\format{
A data frame that mimics ImmuneSpace `con$getDataset("elisa", original_view=TRUE)` results
}
\usage{
sdy1370_elisa
}
\description{
SDY1370 author provided ELISA results formatted to work with ImmuneSpace-derived data
}
\keyword{datasets}
