% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponsePreProcessing.R
\name{selectResponsesToUse}
\alias{selectResponsesToUse}
\title{Generate a single response call data.table from multiple assays}
\usage{
selectResponsesToUse(immdata_age_group)
}
\arguments{
\item{immdata_age_group}{list of assay data.table(s)}
}
\description{
Generate a single response call data.table from multiple assays
}
