% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{subsetToOnlyNeededColumns}
\alias{subsetToOnlyNeededColumns}
\title{Subset to just columns needed for cross-study normalization and batch-correction}
\usage{
subsetToOnlyNeededColumns(geMetaData)
}
\arguments{
\item{geMetaData}{single data.table containing all gene expression meta-data}
}
\description{
Subset to just columns needed for cross-study normalization and batch-correction
}
