% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{summarizeByGeneSymbol}
\alias{summarizeByGeneSymbol}
\title{Summarize probe-level gene expression data by gene symbol selecting the probe
for each gene symbol that has maximum average value across all samples within
that expression set}
\usage{
summarizeByGeneSymbol(esets)
}
\arguments{
\item{esets}{list of expressionSet object}
}
\description{
Summarize probe-level gene expression data by gene symbol selecting the probe
for each gene symbol that has maximum average value across all samples within
that expression set
}
