% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionTests.R
\name{testAllGEMatrixPreNorm}
\alias{testAllGEMatrixPreNorm}
\title{Test gene expression matrix of all samples prior to cross-study normalization}
\usage{
testAllGEMatrixPreNorm(allGE)
}
\arguments{
\item{allGE}{gene expression matrix of all samples}
}
\description{
Test gene expression matrix of all samples prior to cross-study normalization
}
