% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionTests.R
\name{testAllGEMetaDataPreNorm}
\alias{testAllGEMetaDataPreNorm}
\title{Test gene expression meta-data of all samples prior to cross-study normalization}
\usage{
testAllGEMetaDataPreNorm(geMetaData)
}
\arguments{
\item{geMetaData}{gene expression meta-data of all samples}
}
\description{
Test gene expression meta-data of all samples prior to cross-study normalization
}
