% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionTests.R
\name{testExtractedGEData}
\alias{testExtractedGEData}
\title{Test gene expression matrices prior to pre-processing}
\usage{
testExtractedGEData(esets)
}
\arguments{
\item{esets}{list of expressionSet objects}
}
\description{
Test gene expression matrices prior to pre-processing
}
