% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFinalEsets.R
\name{testFinalEset}
\alias{testFinalEset}
\title{Ensure eset is as expected}
\usage{
testFinalEset(eset, hasResponse, ageCohort, ageCutoffs)
}
\arguments{
\item{eset}{expressionSet object}

\item{hasResponse}{whether the eset has response call data}

\item{ageCohort}{whether the eset is for younger cohort}

\item{ageCutoffs}{the age cutoff point for the cohorts}
}
\description{
Ensure eset is as expected
}
