% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionTests.R
\name{testGEMetaDataPreSummarization}
\alias{testGEMetaDataPreSummarization}
\title{Test gene expression meta-data prior to summarization by gene symbol}
\usage{
testGEMetaDataPreSummarization(geMetaData)
}
\arguments{
\item{geMetaData}{gene expression meta-data}
}
\description{
Test gene expression meta-data prior to summarization by gene symbol
}
