% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseTests.R
\name{testImmuneResponseData}
\alias{testImmuneResponseData}
\title{Ensure necessary columns are present in each applicable assay}
\usage{
testImmuneResponseData(selectedImmdata)
}
\arguments{
\item{selectedImmdata}{nested assay data frame lists}
}
\description{
Ensure necessary columns are present in each applicable assay
}
