% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionTests.R
\name{testNoNormEset}
\alias{testNoNormEset}
\title{Test non-normalized base expression set for reasonable gender imputation and
removal of participants without baseline data}
\usage{
testNoNormEset(eset)
}
\arguments{
\item{eset}{non-normalized expression set with all participants}
}
\description{
Test non-normalized base expression set for reasonable gender imputation and
removal of participants without baseline data
}
