% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaccines}
\alias{vaccines}
\title{Vaccine Mapping Data.}
\format{A data frame with seven variables: \code{matrix}, \code{vaccine},
  \code{vaccine_type}, \code{pathogen}, \code{adjuvant}, \code{study}
  and \code{arm_accesion}}
\usage{
vaccines
}
\description{
Mapping of vaccine meta-data to gene expression matrices created by
Thomas Hagan and Evan Henrich
}
\keyword{datasets}
