% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{write_data_metadata}
\alias{write_data_metadata}
\title{Write metadata about data to a csv}
\usage{
write_data_metadata(
  metadata_path,
  dataset_name,
  data_path,
  data = NULL,
  include_counts = FALSE
)
}
\arguments{
\item{metadata_path}{The metadata csv file}

\item{dataset_name}{Name of the dataset}

\item{data_path}{Path to the dataset}

\item{data}{the data}

\item{include_counts}{Include count of subjects, samples and features?
if \code{TRUE}, \code{data} must not be \code{NULL}.}
}
\description{
Write metadata about data to a csv
}
