% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{write_processing_metadata}
\alias{write_processing_metadata}
\title{Write a log of processing date to a csv}
\usage{
write_processing_metadata(metadata_path, task_name)
}
\arguments{
\item{metadata_path}{The metadata csv file}

\item{task_name}{the name of the task}
}
\description{
Write a log of processing date to a csv
}
