% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yChromGenes}
\alias{yChromGenes}
\title{Y-Chromosome Associated Genes}
\format{
string vector containing all gene names
}
\usage{
yChromGenes
}
\description{
Y-Chromosome genes from the ensembl database using the `biomaRt` library
}
\keyword{datasets}
