% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{addCoalescedFeatureSetName}
\alias{addCoalescedFeatureSetName}
\title{Create new featureSetName field that coalesces similar platforms for analysis}
\usage{
addCoalescedFeatureSetName(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Create new featureSetName field that coalesces similar platforms for analysis
}
