% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFinalEsets.R
\name{addResponseData}
\alias{addResponseData}
\title{Add selected immune response data to final expressionSets}
\usage{
addResponseData(eset, immdata)
}
\arguments{
\item{eset}{expressionSets without response data}

\item{immdata}{selected immune response data}
}
\description{
Add selected immune response data to final expressionSets
}
