% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{crossStudyNormalize}
\alias{crossStudyNormalize}
\title{Cross-Study normalize an expressionSet object based on a single
vendor's subset of expression data to define the target distribution}
\usage{
crossStudyNormalize(
  eset,
  targetDistributionVendor,
  targetDistributionExcludedStudies
)
}
\arguments{
\item{eset}{expressionSet}

\item{targetDistributionVendor}{Microarray or RNAseq vendor to use for generating target distribution}

\item{targetDistributionExcludedStudies}{studies to exclude when generating the target distribution}
}
\description{
Cross-Study normalize an expressionSet object based on a single
vendor's subset of expression data to define the target distribution
}
