% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneResponseCallGeneration.R
\name{generateELISAResponse}
\alias{generateELISAResponse}
\title{Generate immune response call for age specific ELISA cohort}
\usage{
generateELISAResponse(dt, discretizationValues, postVaxDayRange)
}
\arguments{
\item{dt}{immdata data.table}

\item{discretizationValues}{points to use for cutting of low, mid, high responders}

\item{postVaxDayRange}{range of allowable study_time_collected}
}
\description{
Generate immune response call for age specific ELISA cohort
}
