% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionCrossStudyNormalization.R
\name{getModelFormula}
\alias{getModelFormula}
\title{Get linear model fit values for each gene}
\usage{
getModelFormula(model.vars)
}
\description{
Get linear model fit values for each gene
}
