% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.genderByMatrix}
\alias{qualityControl.genderByMatrix}
\title{Generate gender box plots (prior to QC)}
\usage{
qualityControl.genderByMatrix(eset)
}
\arguments{
\item{eset}{expressionSet}
}
\description{
Generate gender box plots (prior to QC)
}
