% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControl.samplePlot}
\alias{qualityControl.samplePlot}
\title{Generate sample MDS plots for QC}
\usage{
qualityControl.samplePlot(
  eset,
  method = "MDS",
  numberOfSamples = 10,
  colorCol = "study_accession"
)
}
\arguments{
\item{eset}{expressionSet}

\item{method}{MDS or PCA}

\item{numberOfSamples}{number of samples to use per study}

\item{colorCol}{field name to use for labeling samples in plot}
}
\description{
Generate sample MDS plots for QC
}
