% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFinalEsets.R
\name{testFinalEset}
\alias{testFinalEset}
\title{Ensure eset is as expected}
\usage{
testFinalEset(eset, expectResponse, expectNormalization, ages)
}
\arguments{
\item{eset}{expressionSet object}

\item{expectResponse}{whether the eset has response call data}

\item{expectNormalization}{has the eset been normalized}

\item{ages}{integer vector of expected ages}
}
\description{
Ensure eset is as expected
}
