% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionPreProcessing.R
\name{updateStudyTimepoints}
\alias{updateStudyTimepoints}
\title{Update study_time_collected for known issues}
\usage{
updateStudyTimepoints(geMetaData, studies, originalTp, newTp)
}
\arguments{
\item{geMetaData}{meta-data data.table}

\item{studies}{list of studies to update}

\item{originalTp}{original timepoint value}

\item{newTp}{new timepoint value}
}
\description{
Update study_time_collected for known issues
}
